+++ src/lib/input-validation/validatePermissionResource.ts
@@ -23,7 +23,7 @@ export const validatePermissionResource = (
   value: string,
   { isRequired = true }: ValidatePermissionResourceOptions = {}
 ) => {
-  const REGEX = "^[A-Z]+([:]{1}([A-Z]+|(({[a-zA-Z]+})|\\*)))*$";
+  const REGEX = "^[A-Z]+([:]{1}([A-Z]+|(({[a-zA-Z]+})|[a-zA-Z0-9]+|\\*)))*$";
   if (isEmpty(value)) {
     if (!isRequired) {
       return null;