+++ CHANGELOG.md
@@ -35,4 +35,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added invalidValue in commonValidationErrorType
 - Added validation excluded numbers in validateNumeric
 - Fixed validatePermissionResource regex to allow `*` as variable
+- Fixed validatePermissionResource regex to allow alphanumeric as variable
 - Change some validate functions that has `invalidOption` error, to throw error on `invalidOption` instead of returning it.